refactor(allocator): extract routes/schedules.py from main.py
Moves the scheduled-destruction CRUD (create/get/list/cancel) into its own
blueprint, following the routes/*.py split established in prior tasks.
Hoists the inline status-filter list to a module constant
(VALID_STATUS_FILTERS) and the function-local `datetime` import to module
scope, restoring the normal import arrangement now that nothing shadows it.
